+++ src/index.js
@@ -1,5 +1,5 @@
 /**
- *    Copyright 2020 Greg Lovelidge
+ *    Copyright 2025
 
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 const vscode = require('vscode');
-const { addToDeployment, removeFromDeployment, deployMetadata, viewDeployment } = require('./commands');
+const { addToDeployment,addAllToDeployment, removeFromDeployment, deployMetadata, viewDeployment } = require('./commands');
 const { getSourceFiles } = require('./util');
 let outputChannel;
 /**
@@ -22,9 +22,9 @@ let outputChannel;
  */
 async function activate(context) {
     try {
-        outputChannel = vscode.window.createOutputChannel('Salesforce Deployment Helper');
+        outputChannel = vscode.window.createOutputChannel('Salesforce Deployment Tool');
         // Add to Deployment command
-        let addToDeploymentCmd = vscode.commands.registerCommand('sfdh.addToDeployment', async (sourceUri) => {
+        let addToDeploymentCmd = vscode.commands.registerCommand('sfdt.addToDeployment', async (sourceUri) => {
             // set uri as the active editor there is one
             if (!sourceUri) {
                 const editor = vscode.window.activeTextEditor;
@@ -35,16 +35,36 @@ async function activate(context) {
             const sourceUris = await getSourceFiles(sourceUri, 'Add to Deployment');
             addToDeployment(sourceUris, context, outputChannel);
         });
+        
+        let addAllToDeploymentCmd = vscode.commands.registerCommand('sfdt.addAllToDeployment', async (...commandArgs) => {
 
+            /*let args = {};
+            console.log('test');
+            console.log('files 0 : ', commandArgs[0]);
+            console.log('files 1 : ', commandArgs[1]);
+            if (commandArgs?.length === 1 && !(commandArgs[0] instanceof vscode.Uri)) {   // if from keybinding
+              let argsArray = Object.entries(commandArgs[0]).filter(arg => {
+                return searchCommands.getKeys().includes(arg[0]);
+              });
+              Object.assign(args, Object.fromEntries(argsArray));
+            }*/
+            if (commandArgs[1]?.length > 1) {
+                addAllToDeployment(commandArgs[1], context, outputChannel);
+            }
+            //args.filesToInclude = await parseArgs(commandArgs, "file");
+            //args.triggerSearch = true;
+            //console.log('files 0 : ',args);
+        });
+        
         // Add Multiple Files to Deployment command
-        let addMultipleToDeploymentCmd = vscode.commands.registerCommand('sfdh.addMultipleToDeployment', async () => {
+        let addMultipleToDeploymentCmd = vscode.commands.registerCommand('sfdt.addMultipleToDeployment', async () => {
             const sourceUris = await getSourceFiles(undefined, 'Add to Deployment');
             addToDeployment(sourceUris, context, outputChannel);
         });
 
         // Remove from Deployment command
         let removeFromDeploymentCmd = vscode.commands.registerCommand(
-            'sfdh.removeFromDeployment',
+            'sfdt.removeFromDeployment',
             async (sourceUri) => {
                 const sourceUris = await getSourceFiles(sourceUri, 'Remove from Deployment');
                 removeFromDeployment(sourceUris, context, outputChannel);
@@ -52,23 +72,24 @@ async function activate(context) {
         );
 
         // Deploy command
-        let deployMetadataCmd = vscode.commands.registerCommand('sfdh.deploy', function () {
+        let deployMetadataCmd = vscode.commands.registerCommand('sfdt.deploy', function () {
             deployMetadata(context, outputChannel);
         });
 
         // Clear deployment command
-        let clearDeployMetadataCmd = vscode.commands.registerCommand('sfdh.clearDeployment', function () {
+        let clearDeployMetadataCmd = vscode.commands.registerCommand('sfdt.clearDeployment', function () {
             context.workspaceState.update('deploymentMetadata', []);
             vscode.window.showInformationMessage(`Removed all metadata from the deployment.`);
         });
 
         // View deployment command
-        let viewDeploymentCmd = vscode.commands.registerCommand('sfdh.viewDeployment', () => {
+        let viewDeploymentCmd = vscode.commands.registerCommand('sfdt.viewDeployment', () => {
             viewDeployment(context, outputChannel);
         });
 
         context.subscriptions.push(
             addToDeploymentCmd,
+            addAllToDeploymentCmd,
             addMultipleToDeploymentCmd,
             removeFromDeploymentCmd,
             deployMetadataCmd,
@@ -76,7 +97,7 @@ async function activate(context) {
             viewDeploymentCmd
         );
 
-        vscode.commands.executeCommand('setContext', 'sfdh:project_opened', true);
+        vscode.commands.executeCommand('setContext', 'sfdt:project_opened', true);
     } catch (error) {
         vscode.window.showErrorMessage(`Error! ${error}`);
     }
@@ -85,7 +106,7 @@ exports.activate = activate;
 
 // this method is called when your extension is deactivated
 function deactivate() {
-    vscode.commands.executeCommand('setContext', 'sfdh:project_opened', false);
+    vscode.commands.executeCommand('setContext', 'sfdt:project_opened', false);
     // dispose the channel
     if (outputChannel && outputChannel.dispose) {
         outputChannel.dispose();